Added stats section, turned Globe into server component, removed coordinates endpoint
