+++ app/ui/home/features.tsx
@@ -115,7 +115,7 @@ export default function Features() {
         // preload videos
         <link key={key} rel="preload" as="video" href={demo} />
       ))}
-      <MaxWidthWrapper className="py-10">
+      <MaxWidthWrapper className="pb-10 pt-24">
         <div className="mx-auto max-w-md text-center sm:max-w-xl">
           <h2 className="font-display text-4xl font-extrabold leading-tight text-black sm:text-5xl sm:leading-tight">
             <span className="bg-gradient-to-r from-purple-600 to-pink-600 bg-clip-text text-transparent">
